Add enemy types, bullet capacities, player health
There are now 4 tiers of enemies:
- Type 0: Bounce around screen
- Type 1: Chase player
- Type 2: Chase player and shoot
- Type 3: Avoid player and shoot

Bullets now contain a source property that denotes if the bullet was
fired from an enemy or from the player.

The player can no longer go outside of the boundaries of the game. The
player will also take damage when colliding with an enemy or bullet from
an enemy. Additionally, the player now has a capacity for each type of
bullet, and there is a cap on the number of active bullets (currently
8).

Sounds are no longer scheduled to play and are instead only played as
needed. Support for genres has also been added, but new sound files are
needed.