Add levels to Tempo Attack
This is an update to IGME 330 Project 2. The game is in a playable
prototype phase, with levels, enemy types, and the ability to unlock
additional bullets to add to your inventory when you clear a wave. Also
added is a timer so the user can keep track of their time. For every
user, their best time and level is written to local storage and
displayed.

Skeleton functionality for a leaderboard is also in this commit, but it
has not been fully implemented due to the difficulty of writing to a
file in vanilla JavaScript. This functionality is currently being
replaced with showing the user their highest level and fastest time.